Split example app into oauth and oauth2
